src/third_party/blink/web_tests/http/tests/contacts/select-restricted-to-main-frame.html

<!doctype html>
<meta charset="utf-8">
<title>Contact API: select() restricted to main frame</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<iframe></iframe>

<script>
'use strict';

promise_test(async () => {
  triggerUserGesture();

  var iframe = document.getElementsByTagName('iframe')[0];
  iframe.src = "resources/non-main-frame-select.html";
  return new Promise(function(resolve, reject) {
    window.addEventListener('message', event => resolve(event.data));
  }).then(data => {
    assert_equals(data.errorMsg,
        'TypeError: Unable to open a contact selector');
  });

}, 'Test contacts.select() on a sub-frame')

</script>
